+++ agent/agent.go
@@ -52,7 +52,7 @@ var (
 	registrationFile                     = filepath.Join(appconfig.DefaultDataStorePath, "registration")
 )
 
-func start(log logger.T, instanceIDPtr *string, regionPtr *string) (ssmAgent agent.ISSMAgent, err error) {
+func start(log logger.T, instanceIDPtr *string, regionPtr *string, shouldCheckHibernation bool) (ssmAgent agent.ISSMAgent, err error) {
 	config, err := appconfig.Config(true)
 	if err != nil {
 		log.Debugf("appconfig could not be loaded - %v", err)
@@ -68,9 +68,9 @@ func start(log logger.T, instanceIDPtr *string, regionPtr *string) (ssmAgent age
 	// Do a health check before starting the agent.
 	// Health check would include creating a health module and sending empty health pings to the service.
 	// If response is positive, start the agent, else retry and eventually back off (hibernate/passive mode).
-	if status, err := healthModule.GetAgentState(); status == health.Passive {
+	if status, hibernationErr := healthModule.GetAgentState(); shouldCheckHibernation && status == health.Passive {
 		//Starting hibernate mode
-		context.Log().Info("Entering SSM Agent hibernate - ", err)
+		context.Log().Info("Entering SSM Agent hibernate - ", hibernationErr)
 		go func() {
 			hibernateState.ExecuteHibernation()
 			err = startAgent(ssmAgent, context, log, instanceIDPtr, regionPtr)
@@ -126,7 +126,7 @@ func run(log logger.T) {
 	}()
 
 	// run ssm agent
-	agent, err := start(log, instanceIDPtr, regionPtr)
+	agent, err := start(log, instanceIDPtr, regionPtr, true)
 	if err != nil {
 		log.Errorf("error occurred when starting amazon-ssm-agent: %v", err)
 		return