+++ agent/appconfig/constants_darwin.go
@@ -77,6 +77,8 @@ const (
 	DefaultCustomInventoryFolder = DefaultDataStorePath + "inventory/custom"
 
 	DefaultDocumentWorker = DefaultProgramFolder + "bin/ssm-document-worker"
+	DefaultSessionWorker  = DefaultProgramFolder + "bin/ssm-session-worker"
+	DefaultSessionLogger  = DefaultProgramFolder + "bin/ssm-session-logger"
 
 	// PowerShellPluginCommandName is the path of the powershell.exe to be used by the runPowerShellScript plugin
 	PowerShellPluginCommandName = "/usr/bin/powershell"