+++ agent/contracts/model.go
@@ -192,6 +192,24 @@ type DocumentContent struct {
 	Parameters    map[string]*Parameter    `json:"parameters" yaml:"parameters"`
 }
 
+// SessionInputs stores session configuration
+type SessionInputs struct {
+	S3BucketName                string `json:"s3BucketName" yaml:"s3BucketName"`
+	S3KeyPrefix                 string `json:"s3KeyPrefix" yaml:"s3KeyPrefix"`
+	S3EncryptionEnabled         bool   `json:"s3EncryptionEnabled" yaml:"s3EncryptionEnabled"`
+	CloudWatchLogGroupName      string `json:"cloudWatchLogGroupName" yaml:"cloudWatchLogGroupName"`
+	CloudWatchEncryptionEnabled bool   `json:"cloudWatchEncryptionEnabled" yaml:"cloudWatchEncryptionEnabled"`
+}
+
+// SessionDocumentContent object which represents ssm session content.
+type SessionDocumentContent struct {
+	SchemaVersion string                `json:"schemaVersion" yaml:"schemaVersion"`
+	Description   string                `json:"description" yaml:"description"`
+	SessionType   string                `json:"sessionType" yaml:"sessionType"`
+	Inputs        SessionInputs         `json:"inputs" yaml:"inputs"`
+	Parameters    map[string]*Parameter `json:"parameters" yaml:"parameters"`
+}
+
 // AdditionalInfo section in agent response
 type AdditionalInfo struct {
 	Agent               AgentInfo      `json:"agent"`