+++ agent/framework/coremodules/coremodules.go
@@ -20,6 +20,7 @@ import (
 	"github.com/aws/amazon-ssm-agent/agent/health"
 	"github.com/aws/amazon-ssm-agent/agent/longrunning/manager"
 	"github.com/aws/amazon-ssm-agent/agent/runcommand"
+	"github.com/aws/amazon-ssm-agent/agent/session"
 	"github.com/aws/amazon-ssm-agent/agent/ssm"
 	"github.com/aws/amazon-ssm-agent/agent/startup"
 )
@@ -42,6 +43,10 @@ func RegisteredCoreModules(context context.T) *ModuleRegistry {
 func loadCoreModules(context context.T) {
 	registeredCoreModules = append(registeredCoreModules, health.NewHealthCheck(context, ssm.NewService()))
 	registeredCoreModules = append(registeredCoreModules, runcommand.NewMDSService(context))
+	sessionCoreModule := session.NewSession(context)
+	if sessionCoreModule != nil {
+		registeredCoreModules = append(registeredCoreModules, sessionCoreModule)
+	}
 
 	if offlineProcessor, err := runcommand.NewOfflineService(context); err == nil {
 		registeredCoreModules = append(registeredCoreModules, offlineProcessor)